// backend/world-cup-core/src/basic/model/world-cup-core-error-code.enum.ts
export enum WorldCupCoreErrorCode {
  WC_HISTORICAL_RIVALS_UNAVAILABLE = 'WC_HISTORICAL_RIVALS_UNAVAILABLE',
  WC_SQUAD_UNAVAILABLE = 'WC_SQUAD_UNAVAILABLE',
  WC_TEAM_HISTORY_UNAVAILABLE = 'WC_TEAM_HISTORY_UNAVAILABLE',
  WC_SIMULATION_UNAVAILABLE = 'WC_SIMULATION_UNAVAILABLE',
  WC_SIMULATION_FETCH_FAILED = 'WC_SIMULATION_FETCH_FAILED',
  WC_JOURNEY_UNAVAILABLE = 'WC_JOURNEY_UNAVAILABLE',
  ACTIVE_FINAL_NOT_FOUND = 'ACTIVE_FINAL_NOT_FOUND',
  WC_MATCHES_UNAVAILABLE = 'WC_MATCHES_UNAVAILABLE',
  WC_MATCHES_FETCH_FAILED = 'WC_MATCHES_FETCH_FAILED',
  WC_GROUPS_UNAVAILABLE = 'WC_GROUPS_UNAVAILABLE',
  WC_COACHING_UNAVAILABLE = 'WC_COACHING_UNAVAILABLE',
  WC_TACTICS_UPDATE_FAILED = 'WC_TACTICS_UPDATE_FAILED',
  WC_STATS_UNAVAILABLE = 'WC_STATS_UNAVAILABLE',
  WC_STATS_FETCH_FAILED = 'WC_STATS_FETCH_FAILED',
  WC_WORLD_CUP_HISTORY_UNAVAILABLE = 'WC_WORLD_CUP_HISTORY_UNAVAILABLE',
  BAD_REQUEST = 'BAD_REQUEST',
  SELECTED_STRATEGY_BAD_REQUEST = 'SELECTED_STRATEGY_BAD_REQUEST',
  SELECTED_FORMATION_BAD_REQUEST = 'SELECTED_FORMATION_BAD_REQUEST'
}